GenAI/pyproject.toml
[tool.poetry]
name = "demo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = ">2"
fastapi = "^0.115.3"
google-generativeai = "^0.8.3"
langchain-google-genai = {extras = ["images"], version = "^2.0.1"}
rich = "^13.9.3"
jupyter = "^1.1.1"
pip = "^24.2"
google-cloud-storage = "^2.18.2"
python-multipart = "^0.0.12"
chardet = "^5.2.0"
ruff = "^0.7.1"
pypdf2 = "^3.0.1"
pytesseract = "^0.3.13"
retry = "^0.9.2"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
